+++ integrations/storage/v2/shared/gcs/head.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2024-2025 ArangoDB GmbH, Cologne, Germany
+// Copyright 2025 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -18,24 +18,33 @@
 // Copyright holder is ArangoDB GmbH, Cologne, Germany
 //
 
-package s3
+package gcs
 
 import (
 	"context"
+	"errors"
 
-	"github.com/aws/aws-sdk-go/service/s3"
+	"cloud.google.com/go/storage"
 
-	"github.com/arangodb/kube-arangodb/pkg/util"
+	pbImplStorageV2Shared "github.com/arangodb/kube-arangodb/integrations/storage/v2/shared"
 )
 
-func (i *ios) Delete(ctx context.Context, key string) (bool, error) {
-	_, err := i.client.DeleteObjectWithContext(ctx, &s3.DeleteObjectInput{
-		Key:    util.NewType(i.key(key)),
-		Bucket: util.NewType(i.config.BucketName),
-	})
+func (i *ios) Head(ctx context.Context, key string) (*pbImplStorageV2Shared.Info, error) {
+	b := i.client.Bucket(i.config.BucketName)
+
+	obj := b.Object(i.key(key))
+
+	attr, err := obj.Attrs(ctx)
+
 	if err != nil {
-		return false, err
+		if errors.Is(err, storage.ErrObjectNotExist) {
+			return nil, nil
+		}
+		return nil, err
 	}
 
-	return true, nil
+	return &pbImplStorageV2Shared.Info{
+		Size:          uint64(attr.Size),
+		LastUpdatedAt: attr.Updated,
+	}, nil
 }