+++ integrations/storage/v2/shared/gcs/init.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2024-2025 ArangoDB GmbH, Cologne, Germany
+// Copyright 2025 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -18,24 +18,31 @@
 // Copyright holder is ArangoDB GmbH, Cologne, Germany
 //
 
-package s3
+package gcs
 
 import (
 	"context"
 
-	"github.com/aws/aws-sdk-go/service/s3"
+	"cloud.google.com/go/storage"
 
-	"github.com/arangodb/kube-arangodb/pkg/util"
+	pbImplStorageV2Shared "github.com/arangodb/kube-arangodb/integrations/storage/v2/shared"
+	"github.com/arangodb/kube-arangodb/pkg/util/errors"
 )
 
-func (i *ios) Delete(ctx context.Context, key string) (bool, error) {
-	_, err := i.client.DeleteObjectWithContext(ctx, &s3.DeleteObjectInput{
-		Key:    util.NewType(i.key(key)),
-		Bucket: util.NewType(i.config.BucketName),
-	})
-	if err != nil {
-		return false, err
+func (i *ios) Init(ctx context.Context, opts *pbImplStorageV2Shared.InitOptions) error {
+	if opts.GetCreate() {
+		b := i.client.Bucket(i.config.BucketName)
+
+		if _, err := b.Attrs(ctx); err != nil {
+			if errors.Is(err, storage.ErrBucketNotExist) {
+				if err := b.Create(ctx, i.config.Client.ProjectID, &storage.BucketAttrs{}); err != nil {
+					return err
+				}
+			} else {
+				return err
+			}
+		}
 	}
 
-	return true, nil
+	return nil
 }