+++ pkg/util/constants/constants.go
@@ -61,6 +61,8 @@ const (
 	SecretCredentialsAccessKey = "accessKey" // Key in Secret used to store an S3 API access key
 	SecretCredentialsSecretKey = "secretKey" // Key in Secret used to store an S3 API secret key
 
+	SecretCredentialsServiceAccount = "serviceAccount" // Key in Secret used to store an GCS ServiceAccount File
+
 	SecretAccessPackageYaml = "accessPackage.yaml" // Key in Secret.data used to store a YAML encoded access package
 
 	FinalizerDeplRemoveChildFinalizers = "database.arangodb.com/remove-child-finalizers" // Finalizer added to ArangoDeployment, indicating the need to remove finalizers from all children