+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - (Documentation) Shutdown Debug Feature
 - (Documentation) ManualUpgrade Docs
 - (Documentation) Add Required & Skip in Docs
+- (Feature) (Platform) ECS Storage
 
 ## [1.2.50](https://github.com/arangodb/kube-arangodb/tree/1.2.50) (2025-07-04)
 - (Feature) (Platform) MetaV1 Integration Service